backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
numpy>=1.21.0
pandas>=2.0.0
scikit-learn>=1.3.0
yfinance>=0.2.18
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
aiofiles>=23.2.1
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
fastapi-mail>=1.4.0
email-validator>=2.0.0
